refactor(schema): broke out schema code into multiple files
