+++ packages/schema/src/index.ts
@@ -1,73 +1,14 @@
-import {component, provider, Provider} from '@ziqquratu/core';
-import {Database} from '@ziqquratu/database';
-import {Pipe, PipeFactory, eachDocument} from '@ziqquratu/pipe';
-import Ajv from 'ajv';
+import {component, Provider} from '@ziqquratu/core';
+import {AjvValidator} from './validator';
+import {ValidationConfig} from './interfaces';
+import {ValidationPipeFactory} from './pipe';
 
-export interface ValidationConfig {
-  collection: string;
-}
-
-export class ValidationResult {
-  public constructor(
-    public readonly success: boolean,
-    public readonly errors: Ajv.ErrorObject[]
-  ) {}
-}
-
-export class AjvError extends Error {
-  public constructor(public errors: Ajv.ErrorObject[]) {
-    super(errors[0].message);
-  }
-}
-
-@provider({
-  key: 'schema.Validator',
-  inject: ['ziqquratu.Database', 'schema.ValidationConfig']
-})
-export class Validator {
-  private ajv: Promise<Ajv.Ajv>;
-
-  public constructor(private database: Database, private config: ValidationConfig) {
-    this.ajv = new Promise(resolve => {
-      database.collection(config.collection)
-        .then(collection => collection.find().toArray())
-        .then(docs => resolve(new Ajv().addSchema(docs)));
-    });
-  }
-
-  public async validate(doc: any, schemaId: string): Promise<any> {
-    const validate = (await this.ajv).getSchema(schemaId);
-    if (!validate) {
-      throw new Error('Could not compile schema: ' + schemaId);
-    }
-    if (!validate(doc)) {
-      throw new AjvError(validate.errors || []);
-    }
-    return doc;
-  }
-}
-
-export class ValidationPipeFactory extends PipeFactory {
-  public constructor(private schemaId: string) {
-    super('schema.Validator');
-  }
-
-  public create(): Pipe {
-    return this.resolve((validator: Validator) => {
-      return async (doc: any) => {
-        return validator.validate(doc, this.schemaId);
-      }
-    })
-  }
-}
-
-export const schema = (id: string) => eachDocument(
-  ['insertOne', 'insertMany', 'replaceOne', 'document-upserted'], new ValidationPipeFactory(id),
-);
+export * from './interfaces';
+export {schema} from './pipe';
 
 @component({
   providers: [
-    Validator,
+    AjvValidator,
     Provider.ofInstance<ValidationConfig>('schema.ValidationConfig', {
       collection: 'schemas'
     }),
