+++ packages/schema/src/interfaces.ts
@@ -1,0 +1,7 @@
+export interface ValidationConfig {
+  collection: string;
+}
+
+export interface Validator {
+  validate(doc: any, schemaId: string): Promise<any>;
+}
